Update UploadTicketQuota.cs
Change free_space to long

+++ VimeoDotNet/VimeoDotNet/Models/UploadTicketQuota.cs
@@ -9,7 +9,7 @@ namespace VimeoDotNet.Models
         public bool hd { get; set; }
         public long total_space { get; set; }
         public long space_used { get; set; }
-        public int free_space { get; set; }
+        public long free_space { get; set; }
         public long max_file_size { get; set; }
         public DateTime resets { get; set; }
     }
